+++ packages/app/CHANGELOG.md
@@ -1,5 +1,14 @@
 # app
 
+## 0.0.24
+
+### Patch Changes
+
+- Updated dependencies
+  - @tamagui-extras/form@1.9.0
+  - @tamagui-extras/date@1.8.4
+  - @tamagui-extras/rich-text@1.6.4
+
 ## 0.0.23
 
 ### Patch Changes