+++ packages/date/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @tamagui-extras/date
 
+## 1.8.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @tamagui-extras/form@1.9.0
+
 ## 1.8.3
 
 ### Patch Changes